+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+- add `NumBytes` to `PrimeField`
 - rename package to `@nocturne-xyz/crypto-utils`
 - add module `hybrid-encryption` containing a generic `HybridCipher` 
 - add separate `fromEntropy` method to `PrimeField` for reducing arbitrary `Uint8Array`s into field elements